requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.4
sqlalchemy==2.0.35
pydantic==2.9.2
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
python-telegram-bot==21.7
openpyxl==3.1.5
python-dotenv==1.0.0
PyJWT==2.9.0


